cleaned things up, added basic haddocks
