+++ .gitignore
@@ -6,11 +6,20 @@ bin/
 *.pdf
 *.ps
 *.hp
-dist/
 *-profile
 .hpc/
 *.tix
 *-coverage
 Gemfile.lock
 tmp/
-hellogenetics
+dist
+cabal-dev
+*.chi
+*.chs.h
+*.dyn_o
+*.dyn_hi
+.virtualenv
+.hsenv
+.cabal-sandbox/
+cabal.sandbox.config
+cabal.config
