+++ genetics.cabal
@@ -1,19 +1,38 @@
-Name:           genetics
-Version:        0.0.2
-Category:       Machine Learning
-Synopsis:       A Genetic Algorithm library
-Description:    A Gene typeclass for genetic algorithms
-License:        BSD3
-License-file:   LICENSE.md
-Author:         Andrew Pennebaker
-Maintainer:     andrew.pennebaker@gmail.com
-Build-Type:     Simple
-Cabal-Version:  >=1.6
+name:           genetics
+version:        0.0.3
+category:       Machine Learning
+synopsis:       A Genetic Algorithm library
+description:    A Gene typeclass for genetic algorithms
+license:        BSD3
+license-file:   LICENSE.md
+author:         Andrew Pennebaker
+maintainer:     andrew.pennebaker@gmail.com
+build-type:     Simple
+cabal-version:  >=1.8
 
 source-repository head
   type:     git
-  location: https://github.com/mcandre/genetics
+  location: git://github.com/mcandre/genetics.git
 
-Executable hellogenetics
-    Main-is:        HelloGenetics.hs
-    Build-Depends:  base >= 4.3.1.0 && < 5, random-fu >= 0.1.4 && < 0.3, random-source >= 0.3.0.6, parallel >= 3.2.0.3
+library
+  build-depends:
+    base              >= 4.3.1.0 && <5,
+    random-fu         >= 0.1.4   && <0.3,
+    random-source     >= 0.3.0.6 && <0.4,
+    parallel          >= 3.2.0.3 && <4
+
+  exposed-modules: Genetics
+
+  ghc-options: -Wall
+  hs-source-dirs: src
+
+executable hellogenetics
+  build-depends:
+    base >= 4.3.1.0 && < 5,
+    random-fu,
+    random-source,
+    parallel
+
+  main-is: HelloGenetics.hs
+  hs-source-dirs: src
+  ghc-options: -Wall
